put tests in with .c files
