+++ utils/Matrix/CMakeLists.txt
@@ -1,2 +1,4 @@
 set(CMAKE_C_FLAGS "-Og -g3")
 add_library(MatrixFunctions Matrix.c)
+
+add_library(Matrix_test_lib Matrix_tests.c)
